Day4/requirements.txt
flask
boto3
SQLAlchemy
python-dotenv